requirements.txt
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
psycopg2-binary==2.9.9
